extract mmap_allocator.cpp
This is needed to unit test the way store without dragging
in osm_store.
